requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-filter==21.1
djangorestframework==3.13.0
pytz==2021.3
sqlparse==0.4.2
